requirements.txt
asgiref==3.6.0
attrs==22.2.0
boto3==1.26.49
botocore==1.29.49
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dj-inmemorystorage==2.1.0
dj-rest-auth==2.2.5
Django==4.1.4
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.1
django-storages==1.13.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
exceptiongroup==1.1.0
execnet==1.9.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
moto==4.1.1
null==0.6.1
numpy==1.24.1
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
phonenumberslite==8.13.3
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2
pytest-mock==3.10.0
pytest-xdist==3.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.22.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
svgwrite==1.4.3
toml==0.10.2
tomli==2.0.1
toposort==1.7
types-toml==0.10.8.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
whitenoise==6.2.0
xmltodict==0.13.0
